vscript: move squirrel types to their respective files
