+++ primedev/scripts/scriptjson.cpp
@@ -10,7 +10,7 @@
 #endif
 
 template <ScriptContext context> void
-DecodeJsonArray(HSquirrelVM* sqvm, rapidjson::GenericValue<rapidjson::UTF8<char>, rapidjson::MemoryPoolAllocator<SourceAllocator>>* arr)
+DecodeJsonArray(HSQUIRRELVM sqvm, rapidjson::GenericValue<rapidjson::UTF8<char>, rapidjson::MemoryPoolAllocator<SourceAllocator>>* arr)
 {
 	g_pSquirrel<context>->newarray(sqvm, 0);
 
@@ -49,7 +49,7 @@ DecodeJsonArray(HSquirrelVM* sqvm, rapidjson::GenericValue<rapidjson::UTF8<char>
 }
 
 template <ScriptContext context> void
-DecodeJsonTable(HSquirrelVM* sqvm, rapidjson::GenericValue<rapidjson::UTF8<char>, rapidjson::MemoryPoolAllocator<SourceAllocator>>* obj)
+DecodeJsonTable(HSQUIRRELVM sqvm, rapidjson::GenericValue<rapidjson::UTF8<char>, rapidjson::MemoryPoolAllocator<SourceAllocator>>* obj)
 {
 	g_pSquirrel<context>->newtable(sqvm);
 
@@ -107,7 +107,7 @@ template <ScriptContext context> void EncodeJSONTable(
 {
 	for (int i = 0; i < table->_numOfNodes; i++)
 	{
-		tableNode* node = &table->_nodes[i];
+		SQTable::_HashNode* node = &table->_nodes[i];
 		if (node->key._Type == OT_STRING)
 		{
 			rapidjson::GenericValue<rapidjson::UTF8<char>, rapidjson::MemoryPoolAllocator<SourceAllocator>> newObj(rapidjson::kObjectType);
@@ -240,7 +240,7 @@ ADD_SQFUNC(
 	doc.SetObject();
 
 	// temp until this is just the func parameter type
-	HSquirrelVM* vm = (HSquirrelVM*)sqvm;
+	HSQUIRRELVM vm = (HSQUIRRELVM)sqvm;
 	SQTable* table = vm->_stackOfCurrentFunction[1]._VAL.asTable;
 	EncodeJSONTable<context>(table, &doc, doc.GetAllocator());
