+++ primedev/squirrel/squirrelautobind.h
@@ -15,7 +15,7 @@ extern SquirrelAutoBindContainer* g_pSqAutoBindContainer;
 class __squirrelautobind;
 
 #define ADD_SQFUNC(returnType, funcName, argTypes, helpText, runOnContext)                                                                 \
-	template <ScriptContext context> SQRESULT CONCAT2(Script_, funcName)(HSquirrelVM * sqvm);                                              \
+	template <ScriptContext context> SQRESULT CONCAT2(Script_, funcName)(HSQUIRRELVM sqvm);                                              \
 	namespace                                                                                                                              \
 	{                                                                                                                                      \
 		__squirrelautobind CONCAT2(__squirrelautobind, __LINE__)(                                                                          \
@@ -35,10 +35,10 @@ class __squirrelautobind;
 						returnType, __STR(funcName), argTypes, helpText, CONCAT2(Script_, funcName) < ScriptContext::SERVER >);            \
 			});                                                                                                                            \
 	}                                                                                                                                      \
-	template <ScriptContext context> SQRESULT CONCAT2(Script_, funcName)(HSquirrelVM * sqvm)
+	template <ScriptContext context> SQRESULT CONCAT2(Script_, funcName)(HSQUIRRELVM sqvm)
 
 #define REPLACE_SQFUNC(funcName, runOnContext)                                                                                             \
-	template <ScriptContext context> SQRESULT CONCAT2(Script_, funcName)(HSquirrelVM * sqvm);                                              \
+	template <ScriptContext context> SQRESULT CONCAT2(Script_, funcName)(HSQUIRRELVM sqvm);                                              \
 	namespace                                                                                                                              \
 	{                                                                                                                                      \
 		__squirrelautobind CONCAT2(__squirrelautobind, __LINE__)(                                                                          \
@@ -57,7 +57,7 @@ class __squirrelautobind;
 						__STR(funcName), CONCAT2(Script_, funcName) < ScriptContext::SERVER >);                                            \
 			});                                                                                                                            \
 	}                                                                                                                                      \
-	template <ScriptContext context> SQRESULT CONCAT2(Script_, funcName)(HSquirrelVM * sqvm)
+	template <ScriptContext context> SQRESULT CONCAT2(Script_, funcName)(HSQUIRRELVM sqvm)
 
 class __squirrelautobind
 {
