+++ primedev/vscript/languages/squirrel_re/squirrel/sqlexer.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "vscript/languages/squirrel_re/include/squirrel.h"
+
+// TODO [Fifty]: Verify size
+struct SQLexer
+{
+	unsigned char gap_0[112];
+};
