+++ primedev/vscript/languages/squirrel_re/squirrel/sqvector.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "vscript/languages/squirrel_re/include/squirrel.h"
+
+// TODO [Fifty]: Verify size
+struct SQVector
+{
+	SQObjectType _Type;
+	float x;
+	float y;
+	float z;
+};
